+++ src/search.directive.js
@@ -10,7 +10,8 @@
       scope: {
         'tree': '=',
         'isSearching': '=',
-        'buttons': '='
+        'buttons': '=',
+        'isEditing': '='
       },
       link: function(scope) {
         scope.search = function() {
@@ -18,6 +19,7 @@
             scope.isSearching = scope.query ? true : false;
             scope.tree.setNodesInPath(scope.tree.nodes);
             scope.tree.setCurrentNodes(scope.tree.find(scope.tree.template.title, scope.query));
+            scope.tree.paginate();
           } else {
             scope.clear();
           }
@@ -27,6 +29,7 @@
           scope.query = '';
           scope.isSearching = false;
           scope.tree.setCurrentNodes(scope.tree.getNodesInPath());
+          scope.tree.paginate();
         };
       }
     };