+++ src/sequoia.constants.js
@@ -2,6 +2,7 @@
   'use strict';
 
   var buttons = {
+    root: 'Home',
     edit: 'Edit',
     select: 'Select',
     deselect: 'Deselect',
@@ -14,9 +15,12 @@
     searchClear: '&times;',
     showSelected: 'Show selected',
     hideSelected: 'Hide selected',
+    deselectAll: 'Deselect all',
     backToList: 'Back to list',
     move: 'Move',
-    modalSelect: 'Select'
+    modalSelect: 'Select',
+    up: 'Up a level',
+    searchText: 'Search for an item by title'
   };
 
   var nodeTemplate = {
@@ -27,15 +31,23 @@
 
   var defaultOptions = {
     allowSelect: true,
+    allowSearch: true,
     canEdit: false,
     inline: false,
     buttons: {},
     limit: 0
   };
 
+  var sortableOptions = {
+    sort: true,
+    handle: '.sequoia-move-handle',
+    ghostClass: 'as-sortable-dragging'
+  };
+
   angular.module('ngSequoia')
     .constant('BUTTONS', buttons)
     .constant('NODE_TEMPLATE', nodeTemplate)
-    .constant('DEFAULT_OPTIONS', defaultOptions);
+    .constant('DEFAULT_OPTIONS', defaultOptions)
+    .constant('SORTABLE_OPTIONS', sortableOptions);
 
 })();