+++ src/tree.directive.js
@@ -1,7 +1,7 @@
 (function() {
   'use strict';
 
-  function sequoiaTreeDirective(Tree, BUTTONS, DEFAULT_OPTIONS){
+  function sequoiaTreeDirective(Tree, Utils, BUTTONS, DEFAULT_OPTIONS, SORTABLE_OPTIONS){
 
     return {
       restrict: 'AE',
@@ -9,9 +9,10 @@
       templateUrl: 'angular-sequoia.html',
       scope: {
         'treeNodes': '=sequoiaTree',
-        'model': '=ngModel',
-        'template': '=nodeTemplate',
-        'options': '='
+        'model': '=?ngModel',
+        'template': '=?nodeTemplate',
+        'options': '=?',
+        'path': '=?sequoiaTreePath'
       },
       link: function(scope) {
         function init() {
@@ -21,40 +22,57 @@
           scope.inline = scope.options.inline;
           scope.allowSelect = scope.options.allowSelect;
           scope.isMultiSelect = scope.options.limit === 1 ? false : true;
-          scope.model = scope.isMultiSelect ? (_.isArray(scope.model) ? scope.model : (_.isString(scope.model) ? [scope.model] : [])) : (_.isString(scope.model) ? scope.model : (_.isArray(scope.model) && scope.model.length ? scope.model[0] : ''));
           scope.breadcrumbs = [];
-          scope.tree = new Tree(angular.copy(scope.treeNodes), scope.template);
-          scope.buttons = {
-            edit: scope.options.buttons.edit ? scope.options.buttons.edit : BUTTONS.edit,
-            select: scope.options.buttons.select ? scope.options.buttons.select : BUTTONS.select,
-            deselect: scope.options.buttons.deselect ? scope.options.buttons.deselect : BUTTONS.deselect,
-            goToSubitems: scope.options.buttons.goToSubitems ? scope.options.buttons.goToSubitems : BUTTONS.goToSubitems,
-            addSubitems: scope.options.buttons.addSubitems ? scope.options.buttons.addSubitems : BUTTONS.addSubitems,
-            addNode: scope.options.buttons.addNode ? scope.options.buttons.addNode : BUTTONS.addNode,
-            remove: scope.options.buttons.remove ? scope.options.buttons.remove : BUTTONS.remove,
-            done: scope.options.buttons.done ? scope.options.buttons.done : BUTTONS.done,
-            search: scope.options.buttons.search ? scope.options.buttons.search : BUTTONS.search,
-            searchClear: scope.options.buttons.searchClear ? scope.options.buttons.searchClear : BUTTONS.searchClear,
-            showSelected: scope.options.buttons.showSelected ? scope.options.buttons.showSelected : BUTTONS.showSelected,
-            hideSelected: scope.options.buttons.hideSelected ? scope.options.buttons.hideSelected : BUTTONS.hideSelected,
-            backToList: scope.options.buttons.backToList ? scope.options.buttons.backToList : BUTTONS.backToList,
-            move: scope.options.buttons.move ? scope.options.buttons.move : BUTTONS.move,
-            modalSelect: scope.options.buttons.modalSelect ? scope.options.buttons.modalSelect : BUTTONS.modalSelect
-          };
+          scope.buttons = _.defaults(scope.options.buttons, BUTTONS);
+          scope.sortableOptions = _.assign({}, SORTABLE_OPTIONS, {onSort: handleSort});
+          scope.tree = new Tree(scope.treeNodes, scope.template, scope.buttons);
+
+          scope.model = Utils.setModel(scope.isMultiSelect, scope.model);
+
+          scope.containerStyle = !scope.inline ? { 'overflow': 'scroll', 'max-height': '400px' } : {};
+        }
+
+        function handleSort(evt) {
+          scope.treeNodes = evt.models;
+        }
+
+        function paginate() {
+          scope.tree.paginate();
+          scope.finished = scope.tree.pagination.finished;
         }
 
         scope.load = function(node) {
           scope.onlySelected = false;
-          if(scope.tree.isValidNode(node)) {
-            scope.tree.setCurrentNodes(node[scope.tree.template.nodes]);
-            scope.breadcrumbs = scope.tree.breadcrumbs(node[scope.tree.template.id]);
+
+          var n = node ? node : scope.path ? scope.path : null;
+
+          if(scope.tree.isValidNode(n)) {
+            scope.tree.setCurrentNodes(n[scope.tree.template.nodes]);
+            scope.breadcrumbs = scope.tree.breadcrumbs(n[scope.tree.template.id]);
+            scope.path = n;
+            scope.parentNode = scope.tree.findParentNode(scope.breadcrumbs);
           } else {
             scope.tree.setCurrentNodes();
             scope.breadcrumbs = [];
+            scope.parentNode = null;
+            scope.path = null;
           }
+
+          scope.tree.resetPagination();
+          paginate();
+          scope.$emit('nodes:updated');
+        };
+
+        scope.loadMore = function() {
+          paginate();
         };
 
         scope.select = function(node) {
+          //make sure that scope.model is setup
+          if(_.isUndefined(scope.model)) {
+            scope.model = Utils.setModel(scope.isMultiSelect, scope.model);
+          }
+
           if(node[scope.tree.template.id]) {
             if(scope.options.limit !== 0 && scope.model.length === scope.options.limit) {
               scope.notification = 'You cannot select more than ' + scope.options.limit + ' items!';
@@ -80,6 +98,10 @@
 
         };
 
+        scope.deselectAll = function() {
+          scope.model = scope.isMultiSelect ? [] : '';
+        };
+
         scope.isSelected = function(node) {
           return scope.isMultiSelect ? _.indexOf(scope.model, node[scope.tree.template.id]) !== -1 ? true : false : scope.model === node[scope.tree.template.id];
         };
@@ -94,20 +116,28 @@
             scope.tree.setNodesInPath(scope.tree.nodes);
             scope.tree.setCurrentNodes(selected);
           }
+
+          paginate();
         };
 
         init();
 
-        scope.$watch('treeNodes', function(newVal) {
+        scope.$watchCollection('treeNodes', function(newVal) {
           if(newVal) {
-            scope.tree = new Tree(angular.copy(scope.treeNodes), scope.template);
+            scope.tree = new Tree(scope.treeNodes, scope.template, scope.buttons);
             scope.load();
           }
-        }, true);
+        });
 
-        if(!scope.inline) {
+        scope.$watchCollection('path', function(newVal) {
+          if(newVal) {
+            scope.load();
+          }
+        });
+
+        // if(!scope.inline) {
           scope.load();
-        }
+        // }
 
         /* Handle Modal */
         scope.showModal = function() {
@@ -129,7 +159,7 @@
             return;
           }
 
-          scope.treeNodes = angular.copy(scope.tree.tree);
+          scope.treeNodes = scope.tree.tree;
 
           scope.isEditing = !scope.isEditing;
         };
@@ -139,15 +169,15 @@
             scope.load(node);
           }
 
-          scope.tree.nodes.push(scope.tree.newNode());
+          scope.tree.addNode();
+          paginate();
+
           scope.isEditing = true;
         };
 
         scope.remove = function(node) {
-          var index = node ? _.indexOf(scope.tree.nodes, node) : -1;
-          if(index !== -1) {
-            scope.tree.nodes.splice(index, 1);
-          }
+          scope.tree.removeNode(node);
+          paginate();
         };
 
         scope.closeNotification = function() {
@@ -158,7 +188,7 @@
 
   }
 
-  sequoiaTreeDirective.$inject = ['SequoiaTree', 'BUTTONS', 'DEFAULT_OPTIONS'];
+  sequoiaTreeDirective.$inject = ['SequoiaTree', 'SequoiaTreeUtils', 'BUTTONS', 'DEFAULT_OPTIONS', 'SORTABLE_OPTIONS'];
 
   angular.module('ngSequoia')
     .directive('sequoiaTree', sequoiaTreeDirective);