+++ tests/tree.factory.test.js
@@ -16,13 +16,17 @@ describe('Sequoia Tree Factory', function() {
   beforeEach(module('ngSequoia'));
 
   var SequoiaTree,
+      Utils,
       log,
-      NODE_TEMPLATE;
+      NODE_TEMPLATE,
+      BUTTONS;
 
-  beforeEach(inject(function(_SequoiaTree_, _$log_, _NODE_TEMPLATE_){
+  beforeEach(inject(function(_SequoiaTree_, _SequoiaTreeUtils_, _$log_, _NODE_TEMPLATE_, _BUTTONS_){
     SequoiaTree = _SequoiaTree_;
+    Utils = _SequoiaTreeUtils_;
     log = _$log_;
     NODE_TEMPLATE = _NODE_TEMPLATE_;
+    BUTTONS = _BUTTONS_;
   }));
 
   it('should create a new tree', function() {
@@ -51,11 +55,61 @@ describe('Sequoia Tree Factory', function() {
     expect(tree.isValidNode(node)).toBeTruthy();
   });
 
+  it('should set reset the startkey for pagination and empty the current nodes', function() {
+    var tree = new SequoiaTree(nodes,template);
+
+    tree.setCurrentNodes();
+    expect(tree.nodes).toEqual([]);
+    expect(tree.pagination.startkey).toEqual(0);
+  });
+
   it('should set the current nodes to the tree nodes', function() {
     var tree = new SequoiaTree(nodes,template);
 
     tree.setCurrentNodes();
+    tree.paginate();
     expect(tree.nodes).toEqual(nodes);
+  });
+
+  it('should paginate', function() {
+    var tree, i, _nodes = [];
+
+    _nodes = nodes.concat([0]);
+
+    for(i=0;i<25;i++) {
+      _nodes.push({_id: 'generated_' + i, title: 'Generated title ' + i, nodes: []});
+    }
+
+    tree = new SequoiaTree(_nodes,template);
+
+    tree.setCurrentNodes();
+    tree.paginate();
+    expect(tree.currentNodes.length).toEqual(_nodes.length);
+    expect(tree.nodes.length).toEqual(tree.pagination.limit);
+  });
+
+  it('should not finish pagination', function() {
+    var tree, i, _nodes = [];
+
+    for(i=0;i<25;i++) {
+      _nodes.push({_id: 'generated_' + i, title: 'Generated title ' + i, nodes: []});
+    }
+
+    tree = new SequoiaTree(_nodes,template);
+
+    tree.setCurrentNodes();
+    tree.paginate();
+    expect(tree.pagination.finished).toBe(false);
+  });
+
+  it('should finish pagination', function() {
+    var tree;
+
+    tree = new SequoiaTree(nodes,template);
+
+    tree.setCurrentNodes();
+    tree.paginate();
+    expect(tree.pagination.finished).toBe(true);
   });
 
   it('should set the current nodes to passed nodes', function() {
@@ -63,6 +117,7 @@ describe('Sequoia Tree Factory', function() {
         tree = new SequoiaTree(nodes,template);
 
     tree.setCurrentNodes(someNodes);
+    tree.paginate();
     expect(tree.nodes).toEqual(someNodes);
   });
 
@@ -83,18 +138,25 @@ describe('Sequoia Tree Factory', function() {
 
   it('create breadcrumbs', function() {
     var tree = new SequoiaTree(nodes,template),
-        breadcrumbs = [{ title: 'Root' },{ _id: '9012', title: 'Title three', nodes: [{ _id: '3456', title: 'Subtitle one', nodes: []}, { _id: '7890', title: 'Subtitle two', nodes: [{ _id: '12345', title: 'Subsubtitle one', nodes: []}]}]}, { _id: '7890', title: 'Subtitle two', nodes: [{ _id: '12345', title: 'Subsubtitle one', nodes: []}]}, { _id: '12345', title: 'Subsubtitle one', nodes: []}];
+        breadcrumbs = [{ title: BUTTONS.root },{ _id: '9012', title: 'Title three', nodes: [{ _id: '3456', title: 'Subtitle one', nodes: []}, { _id: '7890', title: 'Subtitle two', nodes: [{ _id: '12345', title: 'Subsubtitle one', nodes: []}]}]}, { _id: '7890', title: 'Subtitle two', nodes: [{ _id: '12345', title: 'Subsubtitle one', nodes: []}]}, { _id: '12345', title: 'Subsubtitle one', nodes: []}];
 
     expect(tree.breadcrumbs('12345')).toEqual(breadcrumbs);
   });
 
+  it('should get the parent node', function() {
+    var tree = new SequoiaTree(nodes,template),
+        path = [{ _id: '9012', title: 'Title', nodes: [] }, { _id: '90122', title: 'Title two', nodes: [] }, { _id: '90123', title: 'Title three', nodes: [] }],
+        node = { _id: '90122', title: 'Title two', nodes: [] };
+
+    expect(tree.findParentNode(path)).toEqual(node);
+  });
+
   it('should build the correct path to a node', function() {
     var tree = new SequoiaTree(nodes,template),
-        path = 'Title three > Subtitle two > Subsubtitle one',
-        found = tree.findSelected(['12345']);
+        testPath = 'Full path: ' + BUTTONS.root + ' » Title three » Subtitle two » Subsubtitle one',
+        path = tree.buildPathToNode({ _id: '12345', title: 'Subsubtitle one', nodes: [] });
 
-    expect(found.length).toBe(1);
-    expect(found[0].title).toBe(path);
+    expect(testPath).toEqual(path);
   });
 
   it('should retrieve the path to a single node passed as a string', function() {
@@ -120,10 +182,23 @@ describe('Sequoia Tree Factory', function() {
   });
 
   it('should create a new empty node', function() {
-    var tree = new SequoiaTree(nodes,template),
-        node = tree.newNode();
+    var tree = new SequoiaTree(nodes,template);
 
-    expect(node).toEqual({_id: node._id, title: '', nodes: []});
+    tree.setCurrentNodes(nodes);
+
+    var node = tree.addNode();
+
+    expect(nodes.length).toEqual(tree.currentNodes.length);
+  });
+
+  it('should remove an existing node', function() {
+    var tree = new SequoiaTree(nodes,template),
+        remove = nodes[0],
+        initialLength = tree.tree.length;
+
+    tree.removeNode(remove);
+
+    expect(initialLength).toEqual(tree.tree.length + 1);
   });
 
 });